feat: add sturcture for albums
