+++ src/albums/interfaces/album.interface.ts
@@ -1,0 +1,8 @@
+interface Album {
+  id: string; // uuid v4
+  name: string;
+  year: number;
+  artistId: string | null; // refers to Artist
+}
+
+export { Album };
